# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Musecat Backend API
  version: 2.0.0
  summary: API v2 wire contract for frontend-facing Musecat endpoints.
  description: |
    API v2 is a breaking cutover: frontend integrations use documented custom endpoints,
    not PocketBase collection REST. See `docs/architecture-contract.md` for authorization,
    state-transition, and migration rules.
servers:
  - url: http://localhost:8090
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitQuery:
      name: limit
      in: query
      description: Positive integer limit. Endpoint-specific caps apply.
      schema:
        type: integer
        minimum: 1
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
        code:
          type: string
...
    SearchResponse:
      type: object
      required: [users, arcades]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/SearchUserItem'
        arcades:
          type: array
          items:
            $ref: '#/components/schemas/SearchArcadeItem'
...
    ArcadeValueResponse:
      type: object
      required: [admin, basic, hour, sns, gtk, game, photo]
      properties:
        admin:
          $ref: '#/components/schemas/ArcadeAdminSummary'
        basic:
          $ref: '#/components/schemas/ArcadeBasicValue'
        hour:
          $ref: '#/components/schemas/ArcadeHourValue'
        sns:
          $ref: '#/components/schemas/ArcadeSnsValue'
        gtk:
          $ref: '#/components/schemas/ArcadeGtkValue'
        game:
          $ref: '#/components/schemas/ArcadeGameValue'
        photo:
          $ref: '#/components/schemas/ArcadePhotoValue'
...
paths:
  /search:
    get:
      tags: [Search]
      summary: /search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Case-insensitive search keyword.
        - $ref: '#/components/parameters/LimitQuery'
        - name: lat
          in: query
          schema:
            type: number
            format: double
          description: User latitude. Provide together with `lon` to boost nearby arcade results.
        - name: lon
          in: query
          schema:
            type: number
            format: double
          description: User longitude. Provide together with `lat` to boost nearby arcade results.
      responses:
        '200':
          description: Search results grouped by resource bucket.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                users:
                  - username: searchhero
                    nickname: Alpha Nick
                    avatar: ""
                arcades:
                  - id: ubbeq9qzfck5s4w
                    country: KR
                    name: Search Palace
                    address: 101 Exact Road
                    nickname: [First Spot]
                    closed: false
                    distance_km: 0.4
        '400':
          description: Missing search query or invalid limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream database lookup failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /arcade:
    get:
      tags: [Arcades]
      summary: /arcade
      description: |
        Returns only a public arcade (open or closed) and relation ids by default.
        Private arcades always return `404`; use authenticated `GET /arcade/draft` for a creator or developer/moderator draft read.
        When `expand=game`, each `game.items[]` entry includes `version`, `series`, `updated`, and the editable game fields.
        Unauthenticated callers receive redacted sensitive data: `tag` is null, `location` and `updated` are empty strings, `price` retains only representative items with `hasHiddenPrices=true` and `accept=null`, and `expand=hour` returns only today's schedule for the arcade timezone.
      parameters:
        - name: id
          in: query
          schema:
            type: string
          description: Arcade id.
        - name: arcade
          in: query
          schema:
            type: string
          description: Alias for `id`.
        - name: expand
          in: query
          schema:
            type: string
          description: |
            `none` or empty returns relation ids only. `all` expands every section.
            Comma-separated values such as `basic,hour` expand only the requested sections.
        - name: source
          in: query
          schema:
            type: string
            enum: [direct, search, nearby, other]
            default: direct
          description: Optional acquisition source recorded with the page view.
        - name: game_series
          in: query
          description: Optional repeated or comma-separated series IDs used to reach the arcade.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Arcade metadata with relation ids or expanded relation payloads.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArcadeValueResponse'
              examples:
                idsOnly:
                  value:
                    admin:
                      id: arcade_123
                      public: true
                      closed: false
                      country: AU
                      timezone: Australia/Sydney
                      createdBy: user_123
                      created: "2026-04-01 12:00:00.000Z"
                      updated: "2026-04-10 09:00:00.000Z"
                    basic: basic_rel
                    hour: hour_rel
                    sns: sns_rel
                    gtk: gtk_rel
                    game: game_rel
                    photo: photo_rel
                expanded:
                  value:
                    admin:
                      id: arcade_123
                      public: true
                      closed: false
                      country: AU
                      timezone: Australia/Sydney
                      createdBy: user_123
                      created: "2026-04-01 12:00:00.000Z"
                      updated: "2026-04-10 09:00:00.000Z"
                    basic:
                      id: basic_rel
                      name: Sydney Arcade
                      address: 123 George St, Sydney NSW
                      direction: Level 2
                      nickname: [Town Hall]
                      subway_line: [T1]
                      location:
                        lat: -33.8708
                        lon: 151.2062
                    hour:
                      id: hour_rel
                      Monday:
                        start: 1000
                        end: 2300
                      Thursday:
                        start: 0
                        end: 0
                      Sunday: 499
                      Note: Late night on Thursdays
                    sns:
                      id: sns_rel
                      items:
                        - type: instagram
                          link: https://instagram.com/example
                          name: Official account
                    gtk:
                      id: gtk_rel
                      items:
                        - type: smoking
                          bool: false
                    game:
                      id: game_rel
                      items:
                        - id: atom_123
                          updated: "2026-04-10 09:00:00.000Z"
                          version:
                            id: version_123
                            series: series_123
                            released_on: "2025-01-01"
                            en: Version A
                            kr: Version A
                            jp: Version A
                            price_default: null
                          series:
                            id: series_123
                            seriesNumber: 1
                            en: Series A
                            kr: Series A
                            jp: Series A
                            en_short: A
                            kr_short: A
                            jp_short: A
                            manufacturer: null
                          location: 2F
                          quantity: 1
                          price:
                            currency: KRW
                            type: custom
                            list:
                              - value: 500
                            accept: [Cash]
                          tag: []
                    photo:
                      id: photo_rel
        '400':
          description: Missing target arcade id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Arcade not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /arcade/analytics:
    get:
      tags: [Arcades]
      summary: /arcade/analytics
      description: |
        Returns cumulative analytics for a public, open arcade. Anonymous and ordinary
        contributors receive only `page_views`, `fault_reports`, and `edit_count`.
        `page_views_by_source`, `series_filter_entries`, `direction_clicks`,
        `visit_verifications`, and `distinct_visitors` are omitted unless the bearer
        user has both the `arcade_owner` tag and the arcade ID in `owns`, or has the
        `developer` or `moderator` tag. The `createdBy` relation alone is insufficient.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: arcade
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Public-safe or privileged cumulative analytics response.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ArcadeAnalyticsResponse'}
        '400':
          description: Missing arcade id.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Arcade does not exist, is private, or is closed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /arcade/analytics/event:
    post:
      tags: [Arcade Mutations]
      summary: /arcade/analytics/event
      description: |
        Records an anonymous direction-click analytics event. The request body is limited to
        2 KiB and accepts at most ten distinct game-series IDs. To limit anonymous metric
        manipulation, each client IP may record at most 20 events per rolling 10 minutes and
        may record an event for the same arcade only once every 30 seconds. Once the arcade is
        validated, rejected series IDs also consume this quota before their database lookup. The server does
        not persist the client IP, user agent, or identity; limits are process-local and retain
        no more than 4,096 active clients.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ArcadeAnalyticsEventRequest'}
      responses:
        '200':
          description: Analytics event recorded.
          content:
            application/json:
              schema:
                type: object
                required: [recorded]
                properties:
                  recorded: {type: boolean, const: true}
        '400':
          description: Invalid event body.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '413':
          description: Analytics event body exceeds 2 KiB.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '429':
          description: Anonymous client exceeded the per-arcade cooldown or rolling event limit. The `Retry-After` header is included.
          headers:
            Retry-After:
              schema: {type: integer, minimum: 1}
              description: Seconds until this anonymous client may retry.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Arcade does not exist or is private.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '502':
          description: Analytics event persistence failure.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /arcade/draft:
    get:
      tags: [Arcades]
      summary: /arcade/draft
      description: Reads a private draft only for its creator or a `developer|moderator`. Public records are intentionally rejected; clients use `/arcade` for public reads.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Arcade metadata with the same shape as `GET /arcade`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ArcadeValueResponse'}
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PocketBaseAuthError'}
        '404':
          description: Arcade does not exist or is not accessible as a draft.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    delete:
      tags: [Arcade Mutations]
      summary: /arcade/draft
      description: Deletes only the authenticated creator's private draft. Published arcade history is never deleted by this endpoint.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Draft deleted.
          content:
            application/json:
              schema:
                type: object
                required: [id, deleted]
                properties:
                  id: {type: string}
                  deleted: {type: boolean}
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PocketBaseAuthError'}
        '403':
          description: Caller is not the draft creator.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Draft not found.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '409':
          description: Published arcades cannot be deleted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /arcade/drafts:
    get:
      tags: [Arcade Mutations]
      summary: /arcade/drafts
      description: Lists the authenticated user's private drafts only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Own draft summaries.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ArcadeDraftListResponse'}
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PocketBaseAuthError'}
  /arcade/memo:
    get:
      tags: [Arcades]
      summary: /arcade/memo
      description: Returns the current immutable memo revision. Public arcades are anonymously readable; private arcades require the creator or a developer/moderator bearer token.
      parameters:
        - name: arcade
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Current memo revision, or null when no memo has been written.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ArcadeMemoResponse'}
        '400':
          description: Missing arcade id.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Arcade does not exist or private access is denied.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    put:
      tags: [Arcade Mutations]
      summary: /arcade/memo
      description: Creates an immutable Tiptap JSON revision and moves the arcade memo pointer. Requires an authenticated user with arcade write access. Saves and changelog creation are atomic.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PutArcadeMemoRequest'}
      responses:
        '200':
          description: Memo saved; unchanged documents return changed=false without creating a revision.
          content:
            application/json:
              schema:
                type: object
                required: [arcade, changed, memo]
                properties:
                  arcade: {type: string}
                  changed: {type: boolean}
                  memo:
                    $ref: '#/components/schemas/ArcadeMemo'
        '400':
          description: Invalid arcade id or Tiptap document.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PocketBaseAuthError'}
        '403':
          description: Authenticated arcade write access is required.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Arcade does not exist or private access is denied.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /arcade/photo/atoms:
    get:
      tags: [Arcade Mutations]
      summary: /arcade/photo/atoms
      description: Lists photo atoms through the custom API. Use each `file_url` for file delivery. Public arcades are available to active contributors; private drafts require creator or developer/moderator access.
      security:
        - bearerAuth: []
      parameters:
        - name: arcade
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Photo atom list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ArcadePhotoAtomListResponse'}
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PocketBaseAuthError'}
        '404':
          description: Arcade does not exist or private draft is inaccessible.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...